+++ examples/ui5-js-app/wdio-webserver.conf.js
@@ -8,8 +8,6 @@ const _config = {
     },
     specs: [join("webapp", "test", "e2e", "**/*.test.js")],
     exclude: [join("webapp", "test", "e2e", "ui5-late.test.js")],
-    logLevel: "error",
-    bail: 0,
     baseUrl: "http://localhost:8888"
 }