+++ src/lib/wdi5-bridge.ts
@@ -2,8 +2,9 @@ import { resolve } from "path"
 import { writeFile } from "fs/promises"
 import { tmpdir } from "os"
 import * as semver from "semver"
+import { mark as marky_mark, stop as marky_stop } from "marky"
 
-import { wdi5Config, wdi5Selector } from "../types/wdi5.types"
+import { clientSide_ui5Response, wdi5Config, wdi5Selector } from "../types/wdi5.types"
 import { WDI5Control } from "./wdi5-control"
 import { WDI5FE } from "./wdi5-fe"
 import { clientSide_injectTools } from "../../client-side-js/injectTools"
@@ -12,8 +13,9 @@ import { clientSide_getSelectorForElement } from "../../client-side-js/getSelect
 import { clientSide__checkForUI5Ready } from "../../client-side-js/_checkForUI5Ready"
 import { clientSide_getUI5Version } from "../../client-side-js/getUI5Version"
 import { clientSide__navTo } from "../../client-side-js/_navTo"
-
+import { clientSide_allControls } from "../../client-side-js/allControls"
 import { Logger as _Logger } from "./Logger"
+
 const Logger = _Logger.getInstance()
 
 /** store the status of initialization */
@@ -40,7 +42,7 @@ export async function setup(config: wdi5Config) {
         browser._controls = []
     }
 
-    addWdi5Commands()
+    _addWdi5Commands()
 
     // inspired by and after staring a long time hard at:
     // https://stackoverflow.com/questions/51635378/keep-object-chainable-using-async-methods
@@ -114,8 +116,7 @@ export async function injectUI5(config: wdi5Config) {
     }
     const waitForUI5Timeout = config.wdi5.waitForUI5Timeout || 15000
     await clientSide_injectTools() // helpers for wdi5 browser scope
-    // expect boolean
-    const result = await clientSide_injectUI5(config, waitForUI5Timeout)
+    const result: boolean = await clientSide_injectUI5(config, waitForUI5Timeout)
 
     if (result) {
         // set when call returns
@@ -135,6 +136,7 @@ export async function checkForUI5Page() {
 }
 
 //******************************************************************************************
+// private
 
 /**
  * creates a string valid as object key from a selector
@@ -186,7 +188,7 @@ function _verifySelector(wdi5Selector: wdi5Selector) {
     return false
 }
 
-export async function addWdi5Commands() {
+async function _addWdi5Commands() {
     browser.addCommand("_asControl", async (wdi5Selector: wdi5Selector) => {
         if (!_verifySelector(wdi5Selector)) {
             return "ERROR: Specified selector is not valid -> abort"
@@ -198,8 +200,34 @@ export async function addWdi5Commands() {
         if (!browser._controls?.[internalKey] || wdi5Selector.forceSelect /* always retrieve control */) {
             Logger.info(`creating internal control with id ${internalKey}`)
             wdi5Selector.wdio_ui5_key = internalKey
-            const wdi5Control = await new WDI5Control().init(wdi5Selector, wdi5Selector.forceSelect)
+
+            marky_mark("retrieveSingleControl")
+
+            const wdi5Control = await new WDI5Control({}).init(wdi5Selector, wdi5Selector.forceSelect)
+
+            const e = marky_stop("retrieveSingleControl")
+            Logger.info(`_asControl() needed ${e.duration} for ${internalKey}`)
+
             browser._controls[internalKey] = wdi5Control
+        } else {
+            Logger.info(`reusing internal control with id ${internalKey}`)
+        }
+        return browser._controls[internalKey]
+    })
+
+    // no fluent API -> no private method
+    browser.addCommand("allControls", async (wdi5Selector: wdi5Selector) => {
+        if (!_verifySelector(wdi5Selector)) {
+            return "ERROR: Specified selector is not valid -> abort"
+        }
+
+        const internalKey = wdi5Selector.wdio_ui5_key || _createWdioUI5KeyFromSelector(wdi5Selector)
+
+        if (!browser._controls?.[internalKey] || wdi5Selector.forceSelect /* always retrieve control */) {
+            wdi5Selector.wdio_ui5_key = internalKey
+            Logger.info(`creating internal controls with id ${internalKey}`)
+            browser._controls[internalKey] = await _allControls(wdi5Selector)
+            return browser._controls[internalKey]
         } else {
             Logger.info(`reusing internal control with id ${internalKey}`)
         }
@@ -215,19 +243,14 @@ export async function addWdi5Commands() {
      * @param {boolean} oOptions.settings.preferViewId
      */
     browser.addCommand("getSelectorForElement", async (oOptions) => {
-        const result = await clientSide_getSelectorForElement(oOptions)
+        const result = (await clientSide_getSelectorForElement(oOptions)) as clientSide_ui5Response
 
-        if (Array.isArray(result)) {
-            if (result[0] === "error") {
-                console.error("ERROR: getSelectorForElement() failed because of: " + result[1])
-                return result[1]
-            } else if (result[0] === "success") {
-                console.log(`SUCCESS: getSelectorForElement() returned:  ${JSON.stringify(result[0])}`)
-                return result[1]
-            }
-        } else {
-            // Guess: was directly returned
-            return result
+        if (result.status === 1) {
+            console.error("ERROR: getSelectorForElement() failed because of: " + result.message)
+            return result.message
+        } else if (result.status === 0) {
+            console.log(`SUCCESS: getSelectorForElement() returned:  ${JSON.stringify(result.result)}`)
+            return result.result
         }
     })
 
@@ -305,6 +328,57 @@ export async function addWdi5Commands() {
 }
 
 /**
+ * retrieve a DOM element via UI5 locator
+ * @param {sap.ui.test.RecordReplay.ControlSelector} controlSelector
+ * @return {[WebdriverIO.Element | String, [aProtoFunctions]]} UI5 control or error message, array of function names of this control
+ */
+async function _allControls(controlSelector = this._controlSelector) {
+    // check whether we have a "by id regex" locator request
+    if (controlSelector.selector.id && typeof controlSelector.selector.id === "object") {
+        // make it a string for serializing into browser-scope and
+        // further processing there
+        controlSelector.selector.id = controlSelector.selector.id.toString()
+    }
+
+    if (
+        typeof controlSelector.selector.properties?.text === "object" &&
+        controlSelector.selector.properties?.text instanceof RegExp
+    ) {
+        // make it a string for serializing into browser-scope and
+        // further processing there
+        controlSelector.selector.properties.text = controlSelector.selector.properties.text.toString()
+    }
+
+    // pre retrive control information
+    const response = (await clientSide_allControls(controlSelector)) as clientSide_ui5Response
+    _writeObjectResultLog(response, "allControls()")
+
+    if (response.status === 0) {
+        const retrievedElements = response.result
+        const resultWDi5Elements = []
+
+        // domElement: domElement, id: id, aProtoFunctions
+        for await (const cControl of retrievedElements) {
+            const oOptions = {
+                controlSelector: controlSelector,
+                wdio_ui5_key: controlSelector.wdio_ui5_key,
+                forceSelect: controlSelector.forceSelect,
+                generatedUI5Methods: cControl.aProtoFunctions,
+                webdriverRepresentation: null,
+                webElement: cControl.domElement,
+                domId: cControl.id
+            }
+
+            resultWDi5Elements.push(new WDI5Control(oOptions))
+        }
+
+        return resultWDi5Elements
+    } else {
+        return "[WDI5] Error: fetch multiple elements failed: " + response.message
+    }
+}
+
+/**
  * can be called to make sure before you access any eg. DOM Node the ui5 framework is done loading
  * @returns {Boolean} if the UI5 page is fully loaded and ready to interact.
  */
@@ -375,17 +449,35 @@ function _getDateString() {
  * @param {Boolean} bReplace
  */
 async function _navTo(sComponentId, sName, oParameters, oComponentTargetInfo, bReplace) {
-    const result = await clientSide__navTo(sComponentId, sName, oParameters, oComponentTargetInfo, bReplace)
-    if (Array.isArray(result)) {
-        if (result[0] === "error") {
-            Logger.error("ERROR: navigation using UI5 router failed because of: " + result[1])
-            return result[1]
-        } else if (result[0] === "success") {
-            Logger.log(`SUCCESS: navigation using UI5 router to hash:  ${JSON.stringify(result[0])}`)
-            return result[1]
-        }
+    const result = (await clientSide__navTo(
+        sComponentId,
+        sName,
+        oParameters,
+        oComponentTargetInfo,
+        bReplace
+    )) as clientSide_ui5Response
+    if (result.status === 1) {
+        Logger.error("ERROR: navigation using UI5 router failed because of: " + result.message)
+        return result.result
+    } else if (result.status === 0) {
+        Logger.log(`SUCCESS: navigation using UI5 router to hash:  ${JSON.stringify(result.status)}`)
+        return result.result
+    }
+}
+
+/**
+ * create log based on the status of result.status
+ * @param {Array} result
+ * @param {*} functionName
+ */
+function _writeObjectResultLog(response: clientSide_ui5Response, functionName: string) {
+    if (response.status > 0) {
+        Logger.error(`call of ${functionName} failed because of: ${response.message}`)
+    } else if (response.status === 0) {
+        Logger.success(
+            `call of function ${functionName} returned: ${JSON.stringify(response.id ? response.id : response.result)}`
+        )
     } else {
-        // Guess: was directly returned
-        return result
+        Logger.warn(`Unknown status: ${functionName} returned: ${JSON.stringify(response.message)}`)
     }
 }