+++ src/types/wdi5.types.ts
@@ -110,6 +110,35 @@ export interface wdi5Selector {
     selector: wdi5ControlSelector
 }
 
+/**
+ * 0 = success
+ * 1 = error
+ */
+export type wdi5StatusCode = 0 | 1
+
+export interface clientSide_ui5Response {
+    status: wdi5StatusCode
+    result?: any // any method
+    message?: string // case of error (status: 1)
+    domElement?: WebdriverIO.Element // getControl
+    id?: string // getControl
+    aProtoFunctions?: Array<string> // getControl
+    className?: string // getControl
+    returnType?: string // executeControlMethod
+    nonCircularResultObject?: any
+}
+
+/**
+ *
+ */
+export interface wdi5ControlMetadata {
+    id?: string // full UI5 control id as it is in DOM
+    methods?: string[] // list of UI5 methods attached to wdi5 control
+    className?: string // UI5 class name
+    $?: Array<string> // list of UwdioI5 methods attached to wdi5 control
+    key?: string // wdio_ui_key
+}
+
 // yet unused
 export interface wdi5Bridge extends Window {
     bridge: RecordReplay